my_blog/requirements.txt
Django==2.2.7
django-password-reset==2.0
Markdown==3.1.1
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.0
django-allauth==0.39.1
django-braces==1.13.0
django-imagekit==4.0.2
django-taggit==0.22.2